Refactor Atom to use a single list of generic particles

+++ radixdlt-kotlin/src/main/kotlin/com/radixdlt/client/core/atoms/Particle.kt
@@ -1,3 +1,3 @@
 package com.radixdlt.client.core.atoms
 
-abstract class Particle internal constructor(private val spin: Int)
+abstract class Particle internal constructor(private val spin: Long)
